SF-2647 Run SMT engine builds in a separate job to the project sync (#2386)
* Stop alternate training sources from being deleted on SMT builds
* Fix crashes caused by missing alternate training corpora and files
* Enqueue the SMT build job separate to the sync job
* Update sync service and sync config to use .Net 8.0 features

+++ src/SIL.XForge.Scripture/Services/MachineProjectService.cs
@@ -679,21 +679,43 @@ public class MachineProjectService(
                 cancellationToken
             );
         }
-        else if (!string.IsNullOrWhiteSpace(alternateTrainingSourceCorpusId))
+        else if (preTranslate && !string.IsNullOrWhiteSpace(alternateTrainingSourceCorpusId))
         {
             // If there is an existing alternate training source
 
             // Remove the corpus from Serval
-            await translationEnginesClient.DeleteCorpusAsync(
-                translationEngineId,
-                alternateTrainingSourceCorpusId,
-                cancellationToken
-            );
+            try
+            {
+                await translationEnginesClient.DeleteCorpusAsync(
+                    translationEngineId,
+                    alternateTrainingSourceCorpusId,
+                    cancellationToken
+                );
+            }
+            catch (ServalApiException e) when (e.StatusCode == StatusCodes.Status404NotFound)
+            {
+                // If the file was already deleted, just log a message
+                string message =
+                    $"Corpus {alternateTrainingSourceCorpusId} in project {buildConfig.ProjectId}"
+                    + " was missing or already deleted.";
+                logger.LogInformation(e, message);
+            }
 
             // Remove the files from Serval
-            foreach (ServalCorpusFile file in oldAlternateTrainingSourceCorpusFiles)
+            foreach (ServalCorpusFile corpusFile in oldAlternateTrainingSourceCorpusFiles)
             {
-                await dataFilesClient.DeleteAsync(file.FileId, cancellationToken);
+                try
+                {
+                    await dataFilesClient.DeleteAsync(corpusFile.FileId, cancellationToken);
+                }
+                catch (ServalApiException e) when (e.StatusCode == StatusCodes.Status404NotFound)
+                {
+                    // If the file was already deleted, just log a message
+                    string message =
+                        $"Corpora file {corpusFile.FileId} for text {corpusFile.TextId} in project {buildConfig.ProjectId}"
+                        + " was missing or already deleted.";
+                    logger.LogInformation(e, message);
+                }
             }
 
             // Remove the reference to the corpus from the project secret
